+++ plutarch-ledger-api/CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `pposixTime` and `unPPosixTime`, to construct and deconstruct `PPosixTime`
   more conveniently
 * `pisEmpty`, `pinclusiveLowerBound`, `pinclusiveUpperBound` to `Interval`
+* `PAssetClass` in `LedgerApi.Value`, parallelling `AssetClass` from
+  `plutus-ledger-api`
 
 ### Changed